added more navbar tests
