+++ app/_components/navbar/Navbar.tsx
@@ -60,7 +60,7 @@ const Navbar = () => {
 
         <Hide below="md">
           {Tags.map((tag) => (
-            <Popover key={tag} trigger="hover" placement="bottom-start">
+            <Popover key={tag} trigger="click" placement="bottom-start">
               <PopoverTrigger>
                 <Text
                   margin={4}
